added example tests
