+++ .gitignore
@@ -2,3 +2,5 @@ target
 output
 log
 *.log
+.classpath
+.settings
